Asks for confirmation when deleting projects.
Finds a better default name for projects.

+++ media/js/controllers/projects-controller.js
@@ -8,14 +8,21 @@ ASTool.ProjectsController = Em.ArrayController.extend(ASTool.RouteBrowseMixin, {
 		},
 
 		deleteProject: function(projectName) {
-			this.get('slyd').deleteProject(projectName);
-			this.removeObject(projectName);
+			if (confirm('Are you sure you want to delete this project? This operation cannot be undone.')) {
+				this.get('slyd').deleteProject(projectName);
+				this.removeObject(projectName);
+			} 
 		},
 
 		createProject: function() {
-			var projectName = 'new_project';
-			this.get('slyd').createProject(projectName).then(function() {
-				this.pushRoute('project', 'Project: ' + projectName, 'fade', projectName);
+			var newProjectName = 'new_project_';
+			var i = 1;
+			while(this.content.any(function(projectName){ return projectName == newProjectName + i })) {
+				i++;
+			}
+			newProjectName += i;			
+			this.get('slyd').createProject(newProjectName).then(function() {
+				this.pushRoute('project', 'Project: ' + newProjectName, 'fade', newProjectName);
 			}.bind(this));
 		}
 	},
